Fix handling of files and switch to node16
As the action fails in outer workflows with
`./services/docker-compose.yml not exists` we try to calculate paths
inside the action path.
We also switch to node `lts/gallium` as this version is mentioned in
official documentation.
A build script was defined in `package.json` to simplify the build of a
new package.
Last but not least we tweak our gitignore a bit to install needed
packages for development but do not commit these to repository.